manejo franquicia retirada

+++ apps/ingredientes/views.py
@@ -4,38 +4,32 @@ from .forms import *
 
 
 def gestionar_ingrediente(request, id_ingrediente=None):
-    """
-    Permite la creación y modificación de ingredientes
-    :param request:
-    :param id_ingrediente:
-    :return:
-    """
-    if id_ingrediente:
-        ingrediente = get_object_or_404(Ingrediente, id=id_ingrediente)
-    else:
-        ingrediente = None
-    form = IngredienteForm(instance=ingrediente)
-    ingredientes = Ingrediente.objects.all()
-    if request.method == 'POST':
-        form = IngredienteForm(request.POST, instance=ingrediente)
-        if form.is_valid():
-            form.save()
-            messages.success(request, 'Ingrediente creado correctamente')
-            return redirect('ingredientes:registrar')
+    if(request.tenant.working==True):
+        if id_ingrediente:
+            ingrediente = get_object_or_404(Ingrediente, id=id_ingrediente)
         else:
-            messages.error(request, 'Por favor verificar los campos en rojo')
-    return render(request, 'ingredientes/gestionar_ingrediente.html', {'form': form, 'ingrediente': ingrediente, 'ingredientes': ingredientes})
+            ingrediente = None
+        form = IngredienteForm(instance=ingrediente)
+        ingredientes = Ingrediente.objects.all()
+        if request.method == 'POST':
+            form = IngredienteForm(request.POST, instance=ingrediente)
+            if form.is_valid():
+                form.save()
+                messages.success(request, 'Ingrediente creado correctamente')
+                return redirect('ingredientes:registrar')
+            else:
+                messages.error(request, 'Por favor verificar los campos en rojo')
+        return render(request, 'ingredientes/gestionar_ingrediente.html', {'form': form, 'ingrediente': ingrediente, 'ingredientes': ingredientes})
+    else:
+        return render(request,"404.html",{})
 
 
 def eliminar_ingrediente(request, id_ingrediente):
-    """
-    Permite la eliminación de ingredientes
-    :param request:
-    :param id_ingrediente:
-    :return:
-    """
-    ingrediente = get_object_or_404(Ingrediente, id=id_ingrediente)
-    ingrediente.delete()
-    messages.success(request, 'Ingrediente eliminado correctamente')
+    if(request.tenant.working==True):
+        ingrediente = get_object_or_404(Ingrediente, id=id_ingrediente)
+        ingrediente.delete()
+        messages.success(request, 'Ingrediente eliminado correctamente')
 
-    return redirect('ingredientes:registrar')
+        return redirect('ingredientes:registrar')
+    else:
+        return render(request,"404.html",{})
